NFDIV-4448: Update TTL using significant events
