+++ src/main/java/uk/gov/hmcts/divorce/solicitor/event/SolicitorSubmitApplication.java
@@ -191,6 +191,7 @@ public class SolicitorSubmitApplication implements CCDConfig<CaseData, State, Us
             .description("Agree statement of truth, pay & submit")
             .showSummary()
             .showEventNotes()
+            .ttlIncrement(36524)
             .showCondition("applicationType=\"soleApplication\" OR [STATE]=\"Applicant2Approved\"")
             .endButtonLabel("Submit Application")
             .aboutToStartCallback(this::aboutToStart)
